+++ src/minterpy/polynomials/utils.py
@@ -56,8 +56,15 @@ def deriv_newt_eval(x: np.ndarray, coefficients: np.ndarray, exponents: np.ndarr
     -----
     - Can compute derivative polynomials without transforming to canonical basis.
     - This derivative evaluation is done by taking derivatives of the Newton monomials.
-    - JIT compilation using Numba was not used here as itertools.combinations() does not work with Numba.
+    - JIT compilation using Numba was not used here as ``combinations()``
+      from the ``itertools`` module does not work with Numba.
+    - Due to multiple nested loops this implementation is very slow except
+      for a small problem (small dimension and small polynomial degree).
 
+    TODO
+    ----
+    - Refactor this initial implementation of polynomial differentiation
+      in the Newton basis.
     """
 
     N, coefficients, m, nr_points, nr_polynomials, x = \